+++ src/dvc_objects/fs/base.py
@@ -474,7 +474,7 @@ class FileSystem:
     ) -> None:
         # Currently, the implementation is non-recursive if the paths are
         # provided as a list, and recursive if it's a single path.
-        from .local import localfs
+        from .implementations.local import localfs
 
         def get_file(rpath, lpath, **kwargs):
             localfs.makedirs(localfs.path.parent(lpath), exist_ok=True)
