+++ src/dvc_objects/fs/generic.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING, List, Optional
 
 from ._callback import DEFAULT_CALLBACK
 from .base import RemoteActionNotImplemented
-from .local import LocalFileSystem
+from .implementations.local import LocalFileSystem
 from .utils import as_atomic
 
 if TYPE_CHECKING:
@@ -47,7 +47,7 @@ def copy(
         return to_fs.put_file(from_path, to_path, callback=callback)
 
     if isinstance(to_fs, LocalFileSystem):
-        from .local import localfs
+        from .implementations.local import localfs
 
         with as_atomic(localfs, to_path, create_parents=True) as tmp_file:
             return from_fs.get_file(from_path, tmp_file, callback=callback)
