+++ src/dvc_objects/fs/implementations/gs.py
@@ -5,7 +5,7 @@ from funcy import cached_property, wrap_prop
 from dvc.scheme import Schemes
 
 # pylint:disable=abstract-method
-from .base import ObjectFileSystem
+from ..base import ObjectFileSystem
 
 
 class GSFileSystem(ObjectFileSystem):
