+++ src/dvc_objects/fs/implementations/hdfs.py
@@ -8,7 +8,7 @@ from funcy import cached_property, wrap_prop
 from dvc.scheme import Schemes
 from dvc.utils import fix_env
 
-from .base import FileSystem
+from ..base import FileSystem
 
 CHECKSUM_REGEX = re.compile(r".*\t.*\t(?P<checksum>.*)")
