+++ src/dvc_objects/fs/implementations/local.py
@@ -8,9 +8,9 @@ from funcy import cached_property, wrap_prop
 
 from dvc.scheme import Schemes
 
-from . import system
-from .base import FileSystem
-from .utils import copyfile, makedirs, move, remove, tmp_fname
+from .. import system
+from ..base import FileSystem
+from ..utils import copyfile, makedirs, move, remove, tmp_fname
 
 logger = logging.getLogger(__name__)
 
@@ -180,7 +180,7 @@ class LocalFileSystem(FileSystem):
 
     @cached_property
     def path(self):
-        from .path import Path
+        from ..path import Path
 
         return Path(self.sep, getcwd=os.getcwd, realpath=os.path.realpath)
